feat(car): 카풀 등록 저장

+++ src/main/java/com/example/napoli/domain/repository/CarpoolRepository.java
@@ -1,7 +1,9 @@
 package com.example.napoli.domain.repository;
 
+
 import com.example.napoli.domain.entity.Carpool;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarpoolRepository extends JpaRepository<Carpool, Long> {
+public interface CarpoolRepository extends JpaRepository<Carpool, Integer> {
+
 }
